requirements-exschool-game.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
jinja2>=3.1,<4.0
itsdangerous>=2.2,<3.0
python-multipart>=0.0.12,<1.0
numpy>=2.1,<3.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
pillow>=10.4,<12.0
pytesseract>=0.3.13,<0.4
scikit-learn>=1.5,<2.0
playwright>=1.55,<2.0
pytest>=8.3,<9.0
